ai_security_lab/requirements.txt
# AI/ML
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2

# Network
requests==2.31.0
dnspython==2.4.2
scapy==2.5.0

# Utilities
pyyaml==6.0
colorama==0.4.6
tqdm==4.65.0
python-dotenv==1.0.0

# Optional (for advanced modules)
# pwntools==4.11.0  # Buffer overflow
# paramiko==3.3.1   # SSH
# cryptography==41.0.0  # Encryption